Add course completion page and logic
Introduces a new route and controller method to display a course completion page when a user finishes a course. Updates the lesson navigator to show a 'Selesai' button on the last lesson, which redirects to the new completion page. The completion page now displays dynamic course data and offers certificate download and navigation to other courses. Also adds logic to check if a user has finished a course in the service layer.